[ts-http-runtime] Port recent Core changes (#28618)
### Packages impacted by this PR

- `@typespec/ts-http-runtime`

### Issues associated with this PR

- Fixes #28614

### Describe the problem that is addressed by this PR

Port of recent changes to Core libraries to the unbranded package. This
should bring the unbranded package back up to date. Once it is up to
date, the plan is to keep it that way by making changes to both Azure
Core and the unbranded core packages simultaneously.

See the PR for more detail on what changes have been ported and a commit-by-commit breakdown.