[identity] MSAL client skeleton (#28493)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

Resolves #28685

### Describe the problem that is addressed by this PR

This PR starts to clarify what a single MsalClient may look like.
There's a lot of existing configuration and such that needed to be hooked up, but I
stuck to the minimal amount needed here. 

The hope is that adding a new credential involves adding a new function
to the MSAL client.

Notable changes:

1. Minimize the amount of state the MsalClient needs to hold. For
example, a clientSecret is not needed for all flows, so it should be a parameter
from ClientSecretCredential#getToken
2. Differentiate between Confidential and Public client applications
3. Use functions instead of inheritance

There is still a lot of open questions, things that need to be plumbed
though, etc. But this is a rough starting point to show what it may look like
and get some feedback.