[CODEOWNERS] Fix issues and label dataplane paths (#28629)
In order to aid with autotriage, we have to set some additional metadata
inside of CODEOWNERS.

The main shift is we need to tag things with #ServiceLabel in order to
have issues be triaged appropriately. Secondarily, for SDKs that are
owned by the SDK team (anyone under Peter) we need to add an
AzureSdkOwners entry even if it's redundant with the path ownership.

Related PRs from other languages:
https://github.com/Azure/azure-sdk-for-net/pull/41995
https://github.com/Azure/azure-sdk-for-java/pull/38797

I didn't touch Mgmt paths, will check with @lirenhe if this is something
he can take care of.

While I was making changes, I cleaned up some errors that GitHub was
complaining about when viewing the file.