Fix the Pre Push Script (#28748)
This PR is not related to any specific package. It is related to the fix
in the pre-push script. In the previous version of the script, there
were 2 issues reported:

1. Script was slow
2. There was a error in the line `err = error.stdout.toString();`.

The second issue is simple and straightforward. There was a missing
check. Once, it is added, the issue is resolved. The first issue is
little tricky.

I always assumed the new packages will be created from the `main` branch
in the `Azure` fork. Based on my recent conversations, it does not seem
to be valid. Some people (based on our project documentation) create the
branch from the `main` branch in their respective fork. (They do sync
their `main` branch with the `Azure` fork frequently. But, this is not
mandatory which causes the problem).As a result, sometimes I see the
difference in about 300+ commits to be validated which causes excess
time.

Now, I have fixed the issue with adding a new remote specific to prepush
tool and using that for comparison. This will reduce the number of
commits significantly which has increased the speed.

@HarshaNalluru Please approve the PR