Adding an environment variable to skip the pre push checks (#28770)
The Pre Push script has some scenarios which is causing issues to
developers. I am working on resolving these issues. But, In the mean
time, there must be an easy way for the developers to push their commits
to github,

As discussed with @xirzec, I am adding an environment variable to skip
these checks.