[identity] Fix nightly tests (#28843)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

Nightly test failures

### Describe the problem that is addressed by this PR

Our nightly tests started failing with a `TypeError: Descriptor for
property generatePluginConfiguration is non-configurable and non-writable` error.

I'm far from an expert here, but I believe the error is due to ESModules
being immutable, whereas CJS Modules are mutable.

Wrapping the stubbable / mockable object is a reasonable workaround to
keep tests green regardless of whether they get run as ESM or CJS

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

Deleting the tests is an option, or an integration test. Neither seem to
fit the bill here. Once we have credentials migrated over we may be able to
delete some of the unit tests and rely on recorded tests to test the various
scenarios. But we are not there yet.