[OpenAI] Edit Customization Layer with Main Typespec (#28533)
### Packages impacted by this PR
@azure/openai

### Issues associated with this PR


### Describe the problem that is addressed by this PR
This PR updates AOAI package to the latest customization tool v2 with
the new folder structure. "sources" folder is removed and the generated
code will be located in "generated" folder. The PR does not use the
latest TypeSpec commit; that will be addressed in a later PR to avoid
massive changes to the API.