[core-rest-pipeline] Address issue with proxy port setting being ignored (#28974)
### Packages impacted by this PR

`@azure/core-rest-pipeline`

### Issues associated with this PR

https://github.com/Azure/azure-sdk-for-js/issues/28951

### Describe the problem that is addressed by this PR

When PR #28563 upgraded `http-proxy-agent` it didn't properly pass along
`port` as this is now encoded in the proxy URL instead of the agent
options.

This PR trims down the unnecessary parsing we are doing to turn the
proxy URL into `ProxySettings`, preferring to pass the URL as-is to
`http-proxy-agent` and avoid losing information. In cases where we
receive `ProxySettings` we will convert it into a URL correctly.

For unknown historical reasons, `getDefaultProxySettings` was exposed as
public surface, but since we no longer need this internally, I have
marked it as deprecated so we may remove it in a later major version.

### Are there test cases added in this PR? _(If not, why?)_

Existing tests were updated to validate the proxy URL being set
correctly on the agent itself.