[identity] Add support for useDefaultBrokerAccount (#28979)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

Resolves #27390

### Describe the problem that is addressed by this PR

Adds `useDefaultBrokerAccount` to `brokerOptions` and plumbs it through
to the InteractiveBrowserCredential. When the user enables this feature,
brokered IBC will first attempt to use the default OS account. If that fails, 
it falls back to the existing behavior.

### Are there test cases added in this PR? _(If not, why?)_

We desperately need to improve our test coverage, but we do not have IBC
tests today.

Co-authored-by: Scott Addie <10702007+scottaddie@users.noreply.github.com>