[identity] Fix default cache name (#29039)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

N/A - noticed in passing

### Describe the problem that is addressed by this PR

When a user does not specify a cache name, we end up stringifying
"undefined" as the cache name. In addition, we prepend a `.` both in 
the constants and the cache name. 

The resulting cache name ends up being `undefined..cae` or
`undefined..nocae` which is not what we want.

This PR fixes both of these issues and matches the MSAL cache name to
the one specified in @azure/identity-cache-persistence.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

Sharing constants between packages may have been better to ensure
consistency; however, that seemed to be overkill for this. Since we _always_ set the
cache name and pass it into the plugin, the reality is that @azure/identity is
the source of truth for how the files are named.