[cosmos] Remove UUID and universal-user-agent (#29064)
### Packages impacted by this PR

- @azure/cosmos

### Issues associated with this PR

- #28955

### Describe the problem that is addressed by this PR

Removes the usage of `universal-user-agent` and `uuid` modules both. The
`universal-user-agent` is ESM and the code footprint is small enough to
rewrite it to internal usage only. The `uuid` module has been replaced
with `@azure/core-util` and the `randomUUID` function.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

Since `universal-user-agent` is ESM only, and cosmos supports both ESM
and CJS, we needed a single solution. We could have used things from
`@azure/core-rest-pipeline` but that's internal, and we don't want to
break existing telemetry.

### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)