Update Get-javascript-DocsMsDevLanguageSpecificPackageInfo in Language-Settings.ps1 (#29075)
This change is necessary to match change in Update-DocsMsMetadata.ps1
made in [this azure-sdk-tools
PR](https://github.com/Azure/azure-sdk-tools/pull/7951). The
Get-javascript-DocsMsDevLanguageSpecificPackageInfo function is the
implementation of the GetDocsMsDevLanguageSpecificPackageInfoFn. This is
something that should have been called all the time but because
javascript was the only language using this, Update-DocsMsMetadata.ps1
was only calling this if the packageInfo file's DevVersion was set.
Because this needs to be called all the time, if it's defined for a
given language, and javascript only needs process if the DevVersion is
set, move that check to the implemented function.