[core-util] Create isNodeLike and isNodeRuntime (#29086)
### Packages impacted by this PR

- @azure/core-util

### Issues associated with this PR

- https://github.com/Azure/azure-sdk-for-js/issues/28101

### Describe the problem that is addressed by this PR

Reverts the behavior of `isNode` checks to ensure that it can still
return true for Bun and Deno. Introducing `isNodeLike` to also have the
same behavior as `isNode` so we can deprecate the former function. We
have also added `isNodeRuntime` which then checks for an explicit
Node.js runtime.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

The previous attempt was to create a `isNodeLike` and be opt-in, versus
this PR which gives the old behavior and new opt-in for strict Node
runtimes.

### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_

- https://github.com/Azure/azure-sdk-for-js/pull/29083

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Deyaaeldeen Almahallawi <dealmaha@microsoft.com>