[@azure/monitor-query]Update Readme file (#29105)
### Packages impacted by this PR
@azure/monitor-query

### Issues associated with this PR
NA

### Describe the problem that is addressed by this PR
The Readme sample is using the incorrect class. This has to be fixed to
the correct class.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
No special design considerations. Just a readme file update

### Are there test cases added in this PR? _(If not, why?)_
No. This is just a readme file change.

### Provide a list of related PRs _(if any)_
NA

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
NA

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

@srnagar Please review and approve the PR.