Fixed error message check in remote rendering tests (#29082)
### Packages impacted by this PR

Only the remote rendering package, but also only the tests, so no
changes customers will see in any capacity.

### Issues associated with this PR

Test failures on the remote rendering pipelines.

### Describe the problem that is addressed by this PR

We changed some error message on the remote rendering service side,
because the old messages were more confusing than helping. This broke
the tests here, because it compared to the old messages still.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

We could revert the error message changes, but that would leave
customers with less helpful information, so I wouldn't do it.

### Are there test cases added in this PR? _(If not, why?)_

No, only changed asserts.