[Test] remove unit tests that verify package version matches (#28653)
These tests were added before we have the azsdk bot automation to
increment version numbers and keep them in sync. They are less useful
now.

I also noticed the app-config test failed to catch the mismatch because
of missing constant entry in package.json. So I changed the test to
verify that version in user-agent header matches sdk version, following
the pattern in the keyvault tests.