[@azure/eventgrid-namespaces] Fix Links in samples (#29318)
### Packages impacted by this PR

@azure/eventgrid-namespaces

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

The `@azure/eventgrid-namespaces` package has been pushed to main branch
last week. For the initial push, the CI did not permit the links to be
pointed to correct location (since the location is not available). This
PR is to fix the links.

Note: I am still waiting for the docs to be published. Once it is done,
I will update those links

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

No special design considerations. Just changing links in Readme files

### Are there test cases added in this PR? _(If not, why?)_

No. Textual changes in Readme file

### Provide a list of related PRs _(if any)_

https://github.com/Azure/azure-sdk-for-js/pull/29245

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

NA

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)