[core-rest-pipeline] Add support for FormData global in Node (#28975)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`

### Issues associated with this PR

- Fixes #28972

### Describe the problem that is addressed by this PR

The global `FormData` is listed as a supported `RequestBodyType`, but is
only supported in browser where `fetch` takes care of it. `FormData` was
previously not available in Node, but since Node 18 it has been
available without a feature This PR makes `formDataPolicy` convert
`FormData` bodies into a `FormDataMap` in Node. Our existing pipeline
then takes care of converting the `FormData` into the appropriate wire
format based on the request content type.

### Are there test cases added in this PR? _(If not, why?)_

Yes, added test cases.

### To do

- [ ] Port to ts-http-runtime