[Test][core] inline build:test commands into browser test scripts
As far as I know for the majority of our packages we don't need to run
`build:test` as we either runs test on .ts files, or `build` script already
generates the .js files under `dist-esm`. Currently we only need to build the
test for browser testing.

This PR moves commands in `build:test` into the browser test scripts, thus
saving us from rebuilding when testing for NodeJS.