[engsys] Update eslint plugin to 6.x (#29505)
### Packages impacted by this PR

`@azure/eslint-plugin-azure-sdk`

### Issues associated with this PR

https://github.com/Azure/azure-sdk-for-js/issues/22698
https://github.com/Azure/azure-sdk-for-js/issues/22699
https://github.com/Azure/azure-sdk-for-js/issues/22700
https://github.com/Azure/azure-sdk-for-js/issues/22701

### Describe the problem that is addressed by this PR

This moves us from v5 -> v6 of eslint-typescript dependencies. The
current stable is 7 with 8 in early alpha. Since major changes include
breaks, it was easier to do this migration one major version at a time,
hence starting with the upgrade to 6. I am hopeful that 6 -> 7 will be
less messy, but I make no promises.

The major annoyance I ran into is you have to create mock "fixtures"
(dummy files) for any test asset paths (e.g. you tell the linter you are
testing `foo/bar/baz.ts` you need a tsconfig somewhere with that file
listed as included *and* you need that exact directory/file structure on
disk somewhere. eslint-typescript blames typescript for this limitation,
but I'm not familiar enough with the parser APIs to know if they're
correct in their analysis.

Note: I haven't gone through to audit all the rules yet, as some are
still a bit busted or need to be removed, but for the sake of updating
our dependencies, I moved everything forward as-is.