[core-rest-pipeline] Improve robustness of tokenCycler
When debugging a recent issue from Storage, it was reported that in the case of the tokenCycler being unable to retrieve a new access token from a CAE challenge, it would fall back to delaying for the maximum lifetime of the current (invalid) access token instead of failing immediately.

This change improves the robustness of this scenario by invaliding the existing cached token when claims are received.