[core] Sanitize URLs in tracing policy (#29606)
### Packages impacted by this PR

@azure/core-rest-pipeline
@typespec/ts-http-runtime

### Issues associated with this PR

Resolves #29433

### Describe the problem that is addressed by this PR

Tracing spans capture the url as an attribute. A customer noticed that
URLs are not sanitized correctly in Java - specifically the QS params.

This was fixed in Java and will now be fixed in JS. It could be
considered a breaking change; however, the OTEL instrumentation 
package is still in beta and as this should have been sanitized earlier 
we can consider this a bugfix.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

I could duplicate the Sanitizer#sanitizeUrl method; however, I wanted to
make sure I can take advantge of the default query params if the list is ever
updated.