[test-credential] Credential relay for browser tests (#29616)
### Packages impacted by this PR

- `@azure-tools/dev-tool`
- `@azure-tools/test-credential`

### Describe the problem that is addressed by this PR

We want to move to user auth and `DefaultAzureCredential` to
authenticate our tests, but `DefaultAzureCredential` does not work in
the browser, blocking this transition.

This PR proposes a solution -- create a short-lived, local-only, server
which can create tokens in Node using DefaultAzureCredential. In the
browser, createTestCredential provides a credential implementation which
calls this server whenever it needs a token, and the server relays the
request to DefaultAzureCredential.

Here's what's changed:

- `test-credential` now returns a credential which requests tokens from
a local web endpoint when running browser tests against a live service.
- `dev-tool` contains an implementation of this web endpoint. The
dev-tool scripts used to run browser tests now start this server
automatically when running browser tests. A separate command to start
the server is also provided for pipelines.
- Update live test pipelines to start the server when running browser
tests.

### Provide a list of related PRs _(if any)_

- Draft version with loads of commits and `/azp run` spam: #29581 
- Harsha's earlier PR to enable DefaultAzureCredential:
https://github.com/Azure/azure-sdk-for-js/pull/29577