[engsys] upgrade dev dependency eslint-plugin-markdown to ^5.0.0
The breaking change that impacts us is that now `typescript` and `javascript`
code blocks in markdown files are maped to virtual files with `.ts` or `.js`,
instead of `.typescript` or `.javascript` as previously in v4.

This PR reacts to the change:

- Add an overrides entry for `**/*.md/*.ts` because they now matches the files
pattern of our first override. Previously they are not linted because of the old
`.typescript` virtual file extension. Their `project` is set to `null` because
they are not part of any project.  They still are benefited from rules that
doesn't require type checking.