[test-util] Move core-tracing to peerDependencies (#29775)
### Packages impacted by this PR

@azure-tools/test-utils

### Issues associated with this PR

Resolves #29736

### Describe the problem that is addressed by this PR

When @azure-tools/test-utils moves to `npm` it will no longer default to
the on-disk version of core-tracing. This impacts our `supportsTracing`
test helper as it no longer sets the instrumenter on the same version as
the client package. Imagine this flow:

<img
src="https://github.com/Azure/azure-sdk-for-js/assets/753570/12d52e8e-16c1-44b3-b57d-edd31a1e4627"
width=70% height=70%>

In this case, test-util and app-configuration are not talking to the
same "module global" instrumenter.

This PR fixes this by removing core-tracing from test-utils, setting it
as a peerDependency instead. All packages that test tracing also
implement tracing, and as such are already depending on core-tracing.

This was verified locally, but I'll need to get it out to NPM before
verification can complete

### Alternatives

See #29709 for a few alternatives:
1. Pass the instrumenter through
2. Use mocks

Both require updates to multiple packages and/or adding public API
surface which is unnecessary. We can fallback to those options if we
have to, but this is a promising alternative.