[amqp messaging] Use zero length byte array for missing message body
Currently we send a data section of `null` when message body is `null` or
`undefined`.  This works fine within JS land but cause interop issues with other
languages because they don't expect a data section of null marker.

This PR changes to use empty byte array to denote body that doesn't exist, which
is the prevailing approach in Amqp space. Internally when data type is `data`,
we send a data section of `Buffer.alloc(0)` to indicate empty body, and decode
`Buffer` of zero byteLength as null on the receiving side.