[Dev Tool] Update allow headers for relay server (#29828)
### Packages impacted by this PR
dev-tool

### Issues associated with this PR


### Describe the problem that is addressed by this PR
We run into preflight CORS issue while testing App Config migration to
DAC in the browser. This PR update the relay server to allow
"Content-Type" and other headers to mitigate the issue.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
The other solution is to use `cors` middleware package to ensure the
preflight request succeeds.