[ServiceBus] Remove verification of abort error messages in tests
After upgrading to `@azure/abort-controller` v2, some tests that verify the
abort error messages have been failing. It turns out that in our async
operations because multiple callbacks are attached to the same abort signal (for
example, retry logic, delay calls for timeout, authentication, auto
lock-renewing, etc.) and multiple abort errors are thrown reacting to their
signal being aborted. After upgrading to `@azure/abort-controller` v2, the order
of `AbortError`s being thrown have changed thus the error we catch in the test
is now different from the one before upgrading. It is possible that our previous
`AbortContoller` implementation of parent-children aborter pattern implies
certain ordering which is different from when manually attaching callbacks.
However, what's important to the consumer of our APIs in this scenario is that
the operation is aborted, AbortErrors are thrown, and one of them is caught,
even a different one than before.

This PR removes verification of the error messages.

I also realize that it is not useful to re-throw the error from delay for
timeout operation, when the delay operation is aborted either because a signal
is aborted, or because a operation it is waiting for succeeds. So that error is
swallowed.