[eslint-plugin] Fix broken global ignore pattern (#29889)
In an ESLint configuration object, if an `ignores` key is used without any other
keys, then the patterns act as global ignores. I added a `name` in my previous
PR which broke it.

This PR fixes the issue by removing the extraneous name. A pattern of
`*.config.{js,cjs,mjs}` is also add to global ignores because we don't need to
lint them.

Also in this PR

- move app-configuration package to use Flat Config format as a test.
- enable cosmosdb test config that was in legacy config format but missed in migration.