[@azure/eventgrid-system-events]Initial Commit V1.0.0-beta.1 (#29688)
### Packages impacted by this PR

@azure/eventgrid-system-events

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

1. Originally, there was one unified package -
[`@azure/eventgrid`](https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/eventgrid/eventgrid).
This package was generated from Swagger specification.
2. Earlier in April 2024, some parts of the `@azure/eventgrid` package
were extracted out to create a new package -
[`@azure/eventgrid-namespaces`](https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/eventgrid/eventgrid-namespaces).
This package was generated from Typespec specification.
3. Now, with this PR, we are looking at the second split of the
`@azure/eventgrid` package. All System Events have been extracted out of
the swagger specification to Typespec specification. This is a new
package named `@azure/eventgrid-system-events`.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

1. In the Typespec core generator, We have an issue where client name
overrides are not applied to `alias`. I have created an issue
https://github.com/microsoft/typespec/issues/3240 to fix this. In the
meantime, I have handled this issue in the custom code.
2. With the new generator, `Known*` models are not generated. I have
created an issue
https://github.com/Azure/autorest.typescript/issues/2476 to fix this. In
the meantime, I have handled this issue in the custom code.
3. In the previous generator, we had several properties which were
overridden as mandatory parameters. With the new generator, such mass
overrides are not possible. So, we are allowing them to be optional and
close to the Wore specification. This is in parity with the other
language SDKs.

### Are there test cases added in this PR? _(If not, why?)_

No. They are all model objects. No code functionality is involved.

### Provide a list of related PRs _(if any)_

1. https://github.com/microsoft/typespec/issues/3240
2. https://github.com/Azure/autorest.typescript/issues/2476
3. https://github.com/Azure/azure-sdk-for-js/pull/29245

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

`npx tsp compile client.tsp`

### Checklists
- [X] Added impacted package name to the issue description
- [X] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@xirzec @joheredi @lmazuel FYI......

---------

Co-authored-by: Harsha Nalluru <sanallur@microsoft.com>