[@azure/monitor-query] [@azure/monitor-ingestion] Update Version Info (#29943)
### Packages impacted by this PR

1. @azure/monitor-query 
2. @azure/monitor-ingestion

### Issues associated with this PR

NA

### Describe the problem that is addressed by this PR

Both the above packages are scheduled for next week release. All the
code changes have been merged. This is to update the version required
based on the prepare release.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

None

### Are there test cases added in this PR? _(If not, why?)_

None

### Provide a list of related PRs _(if any)_

None

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)