[@azure/eventgrid-namespaces]  [@azure/evengrid-system-events] Update Readme file and Cloud Event Model (#29958)
### Packages impacted by this PR

1. @azure/eventgrid-namespaces 
2. @azure/evengrid-system-events 

### Issues associated with this PR

None

### Describe the problem that is addressed by this PR

There is no specific problem associated with this PR. The Readme files
have to be updated for latest changes. This PR accomplishes that. Also,
the casing in he CloudEVent must be modified based on the latest changes
that went in to the Namespaces package.


### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

No design concerns

### Are there test cases added in this PR? _(If not, why?)_

No. Readme file changes do not require any tests.
. 
### Provide a list of related PRs _(if any)_

https://github.com/Azure/azure-sdk-for-js/pull/29764

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Maor Leger <maorleger@users.noreply.github.com>