[@azure/eventgrid-system-events][@azure/eventgrid-namespaces] Rename eventSubscriptionName to subscriptionName, Remove topicName subscriptionName as optional, Update CI YML (#29977)
### Packages impacted by this PR

1. @azure/eventgrid-system-events
2. @azure/eventgrid-namespaces

### Issues associated with this PR

None

### Describe the problem that is addressed by this PR

#### @azure/eventgrid-namespaces

The 2 clients `EventGridSenderClient` & `EventGridReceiverClient` have
the parameters `topicName` and `eventSubscriptionName` as optional. In
order to be consistent with all the languages, these 2 parameters must
be changed to mandatory. Also, the parameter `eventSubscriptionName`
must be changed to `eventSubscriptionName`.

#### @azure/eventgrid-system-events

The `ci.yml` file must be updated to generate the artifact for release. 

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

No special design considerations

### Are there test cases added in this PR? _(If not, why?)_

No test cases are required.

### Provide a list of related PRs _(if any)_

None

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

@maorleger @jeremymeng  Please review and approve the PR