[identity] Define MsalClientOptions (#29999)
### Packages impacted by this PR

@azure/identity

### Issues associated with this PR

Resolves #29888

### Describe the problem that is addressed by this PR

Now that all credentials have been either migrated or a PR opened with
their migration we can narrow down the type of MsalClientOptions to just the
used options.

In the future, we can further refine this by:
- Renaming tokenCredentialOptions to "identityClientOptions"
- Bucket brokerOptions and persistenceOptions under "pluginOptions"

Etc. 

But the goal here is to just have a new definition of used properties so
it can be refined in the future.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

I could do the above recommendations now, but I'd rather keep the scope
small if possible.